requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.20
pyjwt==2.8.0
aioredis==2.0.1
pymysql==1.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
httpx==0.25.0
requests==2.28.2
transformers==4.35.2
torch==2.1.0
redis==5.0.1
passlib[bcrypt]==1.7.4
sentencepiece
protobuf
numpy==1.24.3